Missing a step
After telling us how to define the connection to the source database, you never tell us to click `Connect` at the bottom before selecting the source database. I've added the step to click Connect and renumbered all subsequent steps in that phase of the exercise. There are several other places where you tell us to choose our source database when it is the only one available and already selected. Same with select all tables. It was already selected.